+++ dirty_cat/__init__.py
@@ -11,7 +11,9 @@ from .similarity_encoder import SimilarityEncoder
 from .target_encoder import TargetEncoder
 from .minhash_encoder import MinHashEncoder
 from .gap_encoder import GapEncoder
+from .datetime_encoder import DatetimeEncoder
 from .super_vectorizer import SuperVectorizer
 
+
 __all__ = ['SimilarityEncoder', 'TargetEncoder', 'MinHashEncoder',
-           'GapEncoder', 'SuperVectorizer']
+           'GapEncoder', 'DatetimeEncoder', 'SuperVectorizer']